Prepare IFormulaRewriter factory modification
IFormulaRewriter implementation must not rewrite nodes with a
different factory than the original formula.
As a consequence constructors and methods of IFormulaRewriter
interface and implementors that allow to reference another factory
than formula factory are marked as deprecated. And it now uses the
factory of the original formula to rewrite it.

Add a factory to ITypeCheckingRewriter implementations constructors
and return this factory when needed since it is the correct level of
rewriting to change factory.