Prepare SeqProver for IFormulaRewriter changes
Prepare seqprover code and tests to IFormulaRewriter changes that will
not allow to reference a formula factory since the formulas rewritten
must keep the same factory.

Rewriters are now independent of a formula factory and their
constructor does not take a formula factory as parameter.  Propagate
this change upwards to the callers of the constructor.

Seqprover DLib and Lib are now both static since they use the factory
of the formula rather than keeping an external reference to a factory.

Some static methods of Tactics no longer take a factory argument since
it is now obsolete. UI project is modified consequently.