Prepare SeqProver for IFormulaRewriter changes
Prepare seqprover code and tests to IFormulaRewriter changes that will
not allow to reference a formula factory since the formulas rewritten
must keep the same factory.

Seqprover DLib and Lib are now both static since they must use factory
of the formula used instead of keeping an external reference to a
factory.

The static Tactics.getArithRewriter() method has no longer a factory
argument since the rewriter does not use it. UI project is modified in
consequence for using this method once.