Types are now rewritten and upgraded with formulas
Since types store their factory, the formula ITypeCheckingRewriter
rewriters and upgraders must also rewrite types when they rewrite
formulas to a different factory.

The TypeRewriter now accepts to rewrite the null value for
convenience reason on type rewriting.

A side-effect is that we changed the convention for rewriting an empty
set extension into an empty set.  The new node is not created by the
SetExtension.rewrite() method, but rather by the rewriter itself.  The
advantage of this new solution is that the rewrite method in
SetExtension does not need to know how to rewrite types, and type
rewriting is limited to some implementation of ITypeCheckingRewriter.