Store formula factory used to build a type
A formula factory is now requested for all types construction and
is stored on construction.

The method toExpression() without formula factory as parameter was
added as consequence.

The method toExpression(FormulaFactory) is set as deprecated since
the type contains its own factory.

Set ParametricType constructor as protected to be homogeneous with other
type constructors.

The restriction on formula factory that are documented in the Javaodoc
are not yet enforced.