Enforce factory contract for simple sequents
Since their inception, it was expected that all predicates used to
build a simple sequent would come from the same formula factory. Any
other use would be meaningless. However, we did not have until now the
means to enforce this contract.

Enforce this contract now that we can and document it properly.

As a consequent, the language filter, which is used to filter predicates
of a sequent removing predicate with extensions, is updated to return
predicates built with the default factory, so that the result of
filtering is a simple sequent with the default factory.

Fix corresponding tests which were actually mixing different factories.