Clean IToStringMediator interface and implementors
IToStringMediator do not provide a factory anymore since it must be used
for a particular formula which has a fixed factory.

Store grammar and kind mediator in fields since there are used several
times. Remove factory field that was never used directly.

The forward(Formula) method must be kept with its Formula argument
since we want to print formulas built from the original formula in
particular cases (e.g.: expression extracted from a type of the
formula).

Rename the IToStringMediator.forward(Formula) formula argument from
"child" to "formula" since the argument is either the formula used to
create the mediator or an element that is not a child of the formula.

Fix IToStringMediator clients code that was using the factory that is
not provided anymore.