Ensures that parsing methods use correct version
Since a factory has a particular version, it is forbidden to try to
parse a different version with factory parsing methods. Ensures that is
is not occurring.

As consequence remove language version argument of parseGeneric method
and use the current factory to determine the language version
(precedent inverse behavior was incoherent for a factory instance
method).