Clean V1 to V2 upgrade code
FormulaFactory upgrade methods were not using the target version
argument anymore since there is only a V1 to V2 upgrade available. This
argument is now removed.

Since the upgraders use the factory to upgrade to the resulting version,
we now ensure that the current factory is not the unique V1 factory
(i.e. the factory is V2).

Propagate methods arguments modification and clean test code for
upgrades since the LanguageVersion is now useless to upgrade.