Remove reference to LanguageVersion in BMath classes
Since LanguageVersion is not referenced in the code, remove the language
version field of BMath class. Names of BMathV1 and BMathV2 classes are
sufficient to identify the language version used.

Add toString() methods to BMath children classes that indicate the
language version.

Remove use of BMath.getVersion() in FormulaFactory and use the
toString() method of grammar class instead.