Check factory equality for elements added to type environment
Since the factory equality is enforced in types and formulas, the
factory equality is now also enforced in type environments. A type
environment and its elements are now guaranteed to have the same factory
by raising IllegalArgumentException if an attempt to add an element with
a different factory is done.

Add corresponding tests.