Add factory translation methods for type environments
Since it is forbidden to add elements with a factory different as the
type environment, a translation method and a method to test the
compatibility of a type environment with a given factory were added.

The translation operation is forbidden for inferred type environments
since they are only used for the type checking phase which must not
change the factory.