Add factory translation of type environments
Since it is forbidden to add elements with a factory different as the
type environment, a translation method and a method to test the
compatibility of a type environment with a given factory have been
added.

The translation operation is forbidden for inferred type environments
since they are only used for the type checking phase which must not
change factory.

If we would like to remove the duplicated code testing for the same
factory, we would need to add a type parameter to the TypeEnvironment
class.  It is not worth yet.

Add corresponding tests