Add factory translation of types
Add a Type.translate(FormulaFactory) method which provides the way to
translate a type to a different factory.

Add Type.isTranslatable(FormulaFactory) which tells whethe a type can be
translated to some factory to guarantee that a call to translate() with
this factory will succeed.

The translation operation is forbidden for type variable since they are
only used during the type checking phase which must not change factory.

Add corresponding tests.

Fix a bug in the Type rewriter.