Add factory translation method for types
Add a Type.translate(FormulaFactory) method which provides the way to
translate a type with a different factory.
Add Type.isTranslatable(FormulaFactory) which provides the way to know
if a type can be correctly translated with the given factory to
guarantee that a call to translate() with this factory will succeed.

The translation operation is forbidden for type variable since they
are only used for the type checking phase which must not change of
factory.