Remove kludge for "partition" in lexer
The "partition" operator was installed in all languages, although this
operator has been introduced only since the V2 language.  There was a
kludge in the lexer to repair this after the fact for a V1 lexer.

I have fixed this issue by not making "partition" an operator in the V1
language anymore.  Then, there is no need for this kludge anymore.

This change then showed that the lexer tests were not solid enough.
Indeed, they were using kinds from a V2 grammar, whatever the factory.
But removing the partition operator from the V1 language changes the
number associated to token kinds, hence plenty of test failures.

I have therefore fixed the lexer tests so that they use the grammar
associated with the formula factory when computing kinds.  This also
allowed to remove a similar kludge in these tests.