Enforce element relations when creating elements
Most of the time, element creation is done from a handle.  However,
internal elements can also be created without supplying a handle.
In case the child type is invalid, this would raise an
IllegalArgumentException.  We have to prevent this in an operation and
return a proper error message.

Add corresponding unit tests.