Clean up Rodin EMF tests
Simplify plug-in dependencies (remove hard-coded versions, remove
unused plug-in dependencies).

Remove editing domain tests.  These tests had been added to
investigate using editing domains, which has not be retained in the
Rodin EMF plug-in.

Tagged with @Ignore a performance test that requires the presence of
an external file which is not version controlled.