Implement FR#302 - select mutiple elements in the Rodin Editor
Allow selecting multiple elements with keys and clicks: with the
combination of the second modifier (e.g. SHIFT) and clicks performed
using the the mouse. Selection is homogeneous with regular text editors,
but regular shift+drag events are handled as text selection to select
the textual representation of the models.