Fix NPE during edition of expressions, predicates and assignments
This fix allows to keep a separation in the pretty print between
formulas and the other contributed attributes.
Formulas (i.e. expressions, predicates and assignments) can have various
associated attribute types which are used in the code of the editor.
The solution is to use default types to create damage repairers, and
then create specific content types for each formula intervals as it is
done for attributes.
This is a patch, as one would use information contributed by the editor
element and attribute extensions to configure the rendering of elements
in the editor.