Fix #668 Rodin Editor allows the edition of read-only elements
Now checking that the element is not read-only before letting its
related intervals be edited.

Return null child creation info in case that the parent element is
read-only.

Handlers now check whether concerned elements are read-only or not.
Drag-and-drop is deactivated when the selection contains a child of a
read-only element.