Fix selection when clicking on the visual handle
Sometimes the interval selected by mapping the offset to the list of intervals,
is the interval which ends at the given offset. This fix allows to take into
account the first interval following the one that finishes at the given offset.
This is done by the three following modifications.
Add a method to get the interval following another given interval.
Let the interval return its EditPos
Search for the next interval in the selection controller.