Rework element pasting in Rodin editor
Use the allowed target types to determine if the actual target of a
paste operation is the current element or its parent.

Extract the pasting code to a new class for ease of implementation.

Now inform the user in case of paste or attribute change error, rather
than swallowing the exception.