Fix bug #674 Deadlock while loading buffer
We now check whether we can write the file which has just been upgraded.
If yes, we proceed as before.  If no, we do not attempt to save the
upgraded file (which would cause the deadlock) and just work with an
in-memory copy of it.

In a previous attempt, we tried to implement a deferred write-back of
the in-memory copy, but it was way too difficult to synchronize it with
the rest of the Rodin Database, and we therefore abandoned this idea.