Add dependency on theories for POM
Strangely the computation of the proof status did not depend on the
possible theories used in the project.  This was a mistake as the
language and the rules used in proofs might depend on the version of
the Event-B language.  Therefore, proof statuses must be recomputed
if the used theories change.