Overhaul of datatype parsing in tests
Change the way datatype are specified with a simple string.  The new
version is in my opinion more readable.  It also allows to specify
datatypes where some arguments do not have a corresponding destructor.
Also, the new representation is closer to the internal representation:
the type parameters are not repeated inside a recursive reference of the
type constructor.

Completely re-implement the parser as a consequence.  It is now a very
simple top-down implementation where all parsing is done with regex.