Introduce a proper class for constructor arguments
In the API, there used to be only one class for named arguments, as they
are also destructors, but nothing to describe unnamed arguments.  This
is corrected by this commit which introduces a new class for
representing unnamed arguments.  The destructor class has been
retro-fitted to now derive from the new class.