Move rewriting code to SubstAppli
The code was more dependent on the application than the predicate to
rewrite, therefore it has been extracted to a new method of SubstAppli.

At the same time change the check of the predicate to be replaced: if
it is the wrong predicate, it now raises an assertion failure, rather
than mask the issue.