Fix undo-redo actions and the action update mechanism
Force the undoContext to be null so that Undo and Redo operation are
created as ITextOperationTarget.
Set these actions global to trigger undo-redo operations from the edit
menu.
Modify handlers to expose the isUndo() and isRedo() logic and EMODF
accordingly.
Move the action enablement/disablement into EditorActionUpdater as it is
dedicated to maintain action consistency when the overlay becomes
active.
Add a FocusAdapter to the RodinEditor to update actions, as the
EditorActionUpdater.focusLost() is called before the focus is actually
lost.