Fix history management
Overlay history is now managed by a dedicated undo manager on its
textviewer: removing OverlayEditorUndoManager.
All actions are now redirected to the overlay textviewer when active.
The Rodin editor history manager is never used, and thus removed. This
is done by the RodinConfiguration.getUndoManager() method.
History is updated asynchronously from the focus changes in
the editor. Thus, an HistoryActionUpdater is created to
update the enable state of actions and commands when history
changes occur.
Move some methods to static utility to factorize code.