Fix implicit refreshing of child components editors
Fix the implicit delta processor to filter out delta on the current
root. Indeed implicit children computation does never concern the
current component, but sub-components.
Fix the RecalculateImplicitElementOperation as the implicit adapter is
always on the resource, and not on the root element.
Simplify the ImplicitPresentationUpdater accordingly using the owner
editor as field.