Fix errors by clearing selection
When an element is moved, the database creates a new element at the
specified target and delete the current element. Hence, we can not keep
a reference to the old elements which no longer exist.
The controller selection should be cleared to avoid such case. However,
there might be a better implementation which could maintain the
selection.