Fix Move operation
The move operation is implemented as a copy/delete operation. Thus, a
reference to the element which is moved can not be kept for a further
undo operation. Now update the reference to the moved element at the
target location.
Add a unit test which verifies that the move operation can be undone
and redone when operating on non direct root children.
Add a unit test case to check that renaming of internal names occur
when there would be a name collision during the move operation.