Fix bug #691 Wizards allow modification of generated components [1/2]
Add a method to check if a root is editable (i.e. non read-only and non
generated).
Add a property tester to allow enablement checks when the inputs of the
active editor changes.
Create a definition in org.eventb.ui plug-in to reuse the property
tester in fr.systerel.editor plug-in.