Fix bug #554 : Heterogenous cursors in proving GUI
Fix is divided into two parts : one concerning buttons in the prover UI,
 one concerning hyperlinks.
Add a specific cursor modifier for buttons which tracks mouse movement.
Wraps the checkboxes to add such a cursor modifier as control listener.
Clean up cursor modification algorithm for hyperlinks. Indeed there are
two cases: the cursor is on hyperlink, the cursor is set to a hand
cursor, the cursor is not on hyperlink, the current link is disabled if
any.