Fix resource deltas processing
Move the locking into the BatchOperation so that all deltas are
always processed at the end of the operation.  In the previous code,
some resource-related deltas could get stuck within the delta
processor.

Add a regression test.  Before this fix, two identical deltas were
reported because the resource delta was received too late.