Review and move proof read test after factory change
The test cannotReadSkeletonFactoryChanged in ProofAttemptTests had become wrong
as it is now possible to reuse a proof after a factory change thanks to
factory translation.
Moreover, the test was wrong in the first place: it was only asserting that
a fresh proof attempt with the changed factory was unattempted, which is always
true (the stored proof skeleton with the previous factory was not read).
A new test has been made in TestUserSupports that uses user support capabilities for
loading proof trees from the disk and making new attempts reusing already made proofs.