Introduce dynamic tactic references to handle dynamic tactics
Dynamic tactics used to be serialized in extension.  Thus, the origin of
the dynamic tactic was completely lost once used in a profile, and the
tactic that was applied was the serialized one, which was not the
intended behaviour for dynamic tactics.

Add the IDynamicTacticRef interface, returned by the auto tactic
registry, that contains the id, name and description of a dynamic
tactic.  Instances of this interface are serialized as dynamic tactics,
with the dynamic tactic id alone (instead of the expanded tactic with
combinators, etc).

With this mechanism, dynamic tactic instances are computed when calling
ITacticDescriptor#getTacticInstance(), so as to obtain a fresh instance
from the dynamic tactic provider.

Remove method getDynTactics() which was freezing the tactics at call
time instead of implementing late binding.