Add a combine() method with name and description
The resulting combined tactic returned by ICombinatorDescriptor
used to bear the name and description of the combinator,
which was not relevant for client defined tactics.
It occurred in particular for dynamic tactics, with for instance
'Loop On All Pending' as a tactic name.
Clients can now provide a meaningful name and a description.