Fix form of quantified expressions
The form used to remain implicit when the expression did not contain all
quantified identifiers, or when it contained externally bound identifiers.
The form filter now performs these additional checks, and switches to
explicit form when they are not fulfilled.

Add corresponding tests.

Fix test quantifiedExpression_LambdaToImplicit which was wrong.