Fix bug #705: AssertionError building WD of 'min' with math extensions
Verify validity of names of bound identifiers only with respect to the
default factory to avoid forcing the user to find names that have not
yet been reserved by a mathematical extension.

In quantified formulae, resolve identifiers using the formula factory
of the node instead of the default one.

Add corresponding unit tests.

Remove now unused and dangerous code in QuantifiedUtil.