Require an icon for dynamic dropdown tactic extensions.
Add an icon to the extension for 'Profiles' dropdown.

There used to be only a name, but since we switched to Eclipse 3.8, on linux and windows platforms, the dropdown tool items have space for an image on top of the text, even when no image is set. This make the whole Proof Control toolbar take more vertical space, with dropdown buttons expanding vertically, which is annoying, aesthetically, but most especially because of space loss.
Thus, we now only display the icon, the name is stored but not used for now (like UI tactics in the same extension point).