Fix other conflict in static checker
When a concrete variable conflicts with an abstract parameter, do not
add it to the identifier symbol table, as this will prevent us from
adding the abstract parameter later when processing the event.

As a consequence, we cannot use the filter module anymore but need to do
the filtering directly in the machine variable module.  But the filter
is nevertheless retained for backward compatibility although it does not
do anything anymore.

Add one more test with an extended event which used to break an
assertion of the static checker.