Ensure that variables are checked after context closure
Add a dependency link from the variable module to the context closure
module, because we need that all identifiers from contexts are present
when checking variable name conflicts.

Add a test where a variable conflicts with a carrier set coming from
a context seen by the abstract machine but not the concrete machine
(and therefore copied implicitly).