Add tryPutSymbolInfo() in symbol tables
The Static Checker code relies on some CoreException being raised by
symbol tables.  This is a very bad idea.

However, we cannot change this exception without breaking the
API.  Therefore, we instead add a new method which performs the same
action, but returns a boolean to indicate success.

Enhance unit tests to check this new implementation.