Fix bug #713: GenMP makes the prover loop
Define a dedicated method in ProverSequent for the Rewrite Hypothesis
action and call it from the action.

Factorize common code within ProverSequent between forward inference and
rewrite. Rewrite needs to be performed entirely as soon as it is
applicable. In particular, the hide action must be executed even if the
forward inference does not modify the sequent.

Add unit test on Rewrite Hypothesis Action.

Add tests on the Generalized Modus Ponens corresponding to the example
in the bug report.