Use marker matchers for SC predicate tests
Revise all the predicate tests of the Static Checker to use marker
matchers without changing the tests.  This allows to have a more
complete view of the markers that are generated by the Static Checker.

This has needed to revise the interface for generic tests in order to
pass the expected IRodinProblem for label conflicts.