Remove test suite 'AllTests' in AST tests
This test suite was annoying as, there were often missing classes.
Now letting JUnit and Maven discover test classes.
This required an adaptation of the POM so as to include inner classes.
Two of the inner classes had a superfluous '@Test' annotation that was
causing Maven to complain about their constructor. It has been removed.