Protect against load and save factory error
Prune the root node in case of formula factory storage problems.

In case of formula factory loading problems, the formula factory of the
proof component is used to try to re-build the proof tree. The rebuild may fail
by returning null or throwing a core exception.

Add corresponding unit tests.