Mostly fix the concurrency issue in the Event-B Explorer
Now performs the changes to the model in the UI thread.  This should
allow to avoid most concurrency issues, at the potential cost of a less
responsive UI.  The issues which are not fixed are the case where the
model is read outside of the UI thread, but this should happen less
frequently.