Bug #721: Denormalized hypothesis not hidden
Change the way forward inference actions are applied to a sequent so
that already existing hypotheses do not prevent hiding rewritten
hypotheses.

This is a non-compatible change for the automated rewriters, but I think
we can live with it, as it fixes a very annoying bug.

Add corresponding regression tests showing that the new behavior is
sensible.