Fix deadlock in Proof Skeleton update
Fix a concurrency issue by running the update asynchronously.  Indeed,
the update needs to lock parts of the database to read the proof and
this would block if the proof has already been locked (which is the case
most of the time where the update is triggered by a change in the proof).