Bug #726: Wrong factory after specialization
Fixes two bugs about factory translation in specialization: one for
specializing given types (as in the SourceForge bug report) and another
one when specializing type environments.

Complete all tests about specialization to also test the case where the
source and destination factories are different and showing that the
specialized formulas are indeed built with the target factory.