Bug #727: Type specialization incomplete
When translating a type, the specialization did not produce any
side-effect.  However, when encountering a given type without any
attached substitution, it would specialize it to itself, which need to
be remembered to ensure consistence of repeated application of the same
specialization.

This commit adds the necessary side-effect and adds a regression test
showing that the bug is fixed.