Add datatype origin.
Formula extensions provide an origin object, that is usually used by
SC modules to keep trace of the DB element that defines them.
Concerning datatype related extensions, the origin is set to the datatype.
So it's not possible to get back to the defining DB element.
In order to fix this, an origin can now be passed to the datatype builder,
ending up in the finalized datatype.

Also consider origin in datatype equality, for the datatype cache
to distinguish two similar datatypes with a different origin (thus
potentially a different mathematical context).

Also add tests to ensure that the correct origin is found in the
finalized datatype.