Add a proof rebuild method with tryReplayUncertain parameter
This parameter can be used to distinguish cases where we want to
merely display a proof as is, although it may not be directly reusable,
from cases where we want to get rid of uncertain rules as much as
possible.
This new method also changes previous behaviour that used to fail
when a rule was not reusable: now rules are always applied to nodes.