Fix bug #729: misleading error message
When a predicate or expression is expected in the input area of the
Proof Control view, print a proper error message, rather than the
generic "Parse error for".

Note that MultiExprInput corresponds to quantifier instantiation and
that it is authorized to have only a partial instantiation, i.e., some
expressions missing. Therefore this class has not been changed.