Fix combo box edit
As reported by Andy on the developer mailing list, combo boxes
used to be reset to --undef-- value right after selecting another
combo box, so the previously selected value was just discarded.
Now using setItems() and forcing the contents of the box to be
a member of the list of items.