Fix bug #742 Rodin Editor loses selection when moving
Modify the implementation of the atomical move operation to avoid
automatic renaming of internal elements which leads to selection loss.
Renaming now solely occurs when there could be a naming collision. In
the implemented editor, this case is limited to drag operations when
target element parent differs from source element parent. Selection
would indeed be lost in that specific case, but this behavior is
accepted.