Fix the root cause of failing test in ContrHypsTests
Son had modified this test, I suppose because it was failing on its
computer.  What happens is that the original test was working on
Java 6 and Java 7, but failing with Java 8.  This is because the test
was relying on the order of elements in a HashSet.

I have changed the test code so that it is now independent from the
underlying implementation of HashSet.