Port TestEqualitySolver to Java 8
This test was relying on the order in which a HashSet is traversed,
which introduced a regression in Java 8 (as this algorithm which is
unspecified has changed).

Do not rely anymore on this order: sort the elements of the set to
obtain a consistent order, whatever the implementation of HashSet.