Maintain order of variables in newPP
When copying a clause, the order of variables could change which made
the result of resolution unpredictable (it was depending on the order
in which a HashSet would be traversed).  The order is now maintained
through the copy.

Update TestResolution to take into account the preserved order: local
variables are now numbered in the reverse order (evar0 and evar1 are
swapped) in testRule2.

All tests now pass with Java 6, 7 or 8.