Ensure that images of prover sequents are predictable
The toString() method of ProverSequent used to call the toString()
method of its type environment.  However, this latter method calls
toString() on a HashMap, which gives an unpredictable output.

One option would have been to change the toString() method of
type environments, but I am not sure of the consequences.  I have
therefore only changed the toString() method of ProverSequent so that
all the org.eventb.core.tests pass with all supported versions of Java
(6, 7 and 8).

Updated the POLoaderTests accordingly (sorting the elements of the
type environments of expected sequents).