Fix order dependent Rodin Builder tests
Some tests of the Rodin Builder needed to be run before others.  Namely,
class CBuilderTests needed to be run before class MBuilderTests.  This
was due to the missing initialisation of trace variables.

This commit adds the necessary setup for all Builder tests so that these
variables are always initialized with a default value.

The order in which these tests are run is now irrelevant.