Fix bug Theory #8
When a simple sequent contains a formula extension which is not
WD-strict, we cannot translate it to the plain mathematical language.

What we do instead is to remove any predicate that contains such an
operator, and then to translate the remaining extensions.

This gives a stronger sequent, but at least we can expect an external
prover to discharge it if it does not depend on the semantics of the
removed/translated extensions.

Add corresponding unit tests.