Translate properly extensions in type environment
For the target type environment, we now need to translate it in two
phases.  We first create an empty one, then we apply the typeRewriter
to each type occurring in the type environment (this will have the side
effect of adding the translation of extension types to the environment).

As we are using a separate list of used identifiers, we know that
whatever the order in which identifiers are added to the environment,
we will get the same result.