Implement translation of parametric types with parameters
The type constructor is translated as a relation, as was already done
for datatype type constructors.  This comes naturally as the type
constructor takes sets and produces a set, which is exactly the
signature of a relational application.

Add several tests to validate that the translation is sound with several
kinds of parametric types.