Simplify final tests of canPut methods
For types, streamline the test and avoid an if for returning true or
false, return directly the Boolean expression.

For expressions, do not compare the substitutes, but rather the value
with the one that can be retrieved from the substitution.  This is
more neat and avoids creating an additional Substitute object.

No functional change.