Resurrect test about new Specialization getters
The test was missing the @Test annotation and was therefore never run.

When resurrecting it, I found it failing for several reasons which
have been fixed.  I have also defined a dedicated assertion method
for checking that the returned array is OK, irrespective of the
order of the elements (they come from a set).

Also, the original test was adding twice the same identifier, but with
different types. This will not be allowed anymore, so rename the second
entered identifier.

No functional change.