Improve canPut for identifiers
Rather than having the verify method throw an exception and catch it
later, it is better to have this method return a Boolean value.  Note
that “speTypeRewriter.rewrite(type)” never throws an exception here.

There is still the issue of the rewrite() method creating new identity
substitution in the speTypeRewriter which will be fixed later.

No functional change.