Avoid side-effects for incompatible ident substitutions
When checking with canPut() or adding with put() a new identifier
substitution, if the type-check fails, then no identity type
substitution shall be added.

For canPut(), even if type-check succeeds, there shall not be any
side-effect.

Add corresponding tests.  Note that there is no need to add tests when
the type-check succeeds, as this was already tested.

Now all the canPut() methods are side-effect free.