Implement side-effect for predicate variable substitution
This was not implemented, which makes predicate variable substitution
behave differently from type and identifier substitution. Not
implementing it means that after specialization we could have some
predicate variables that have been substituted only partially, depending
on the relative order of calls to specialize() and put().

Add corresponding documentation and unit test.