Add tests about failing specialization
When one attempts to specialize a type, type environment or formula
which is incompatible with the specialization (differing source type
environments), the specialization shall fail.

Also add tests ensuring that we do not have false positives, when the
potentially incompatible identifier gets substituted and thus
disappears.

Complete the Javadoc of the specialize() methods to make this
restriction explicit.

No functional change.