Load the B font asynchronously
Previously, the B font, if not yet available, was loaded during the
startup of the org.eventb.ui plugin.  But this was introducing deadlocks
when running UI tests which got resolved after a 5 second delay. As
the deadlock was reappearing several times, it was causing a total delay
of more than 2 min 30 sec. It could also introduce random crashes when
running the Event-B UI tests.

With the asynchronous loading, the tests are now running normally.