Move BundledFileExtractor to Rodin keyboard UI
Later, the installation of the math font will be performed by the
Rodin keyboard plugin (where it should have been done to start with).
We prepare this change by moving the file extractor to the Rodin
keyboard plug-in and publish it so that it can be shared by all UI
plugins.