Rework details: use a StackLayout
Use a StackLayout for the detail container.
The StackLayout corresponds to the intention of having several
composites at the same place, with one on top of the others.
The previous management with setVisible was making the composites
disappearing upon resizing.

Also fix parameterized profile viewer to show the tactic name,
by adding a grid parent of the parameterized components (tactic name and table).