Rodin editor: fix updating of removed attributes
The retainAll call was passed a collection of String,
while the method would only work with collections of
Map.Entry<String, Attribute>, so the list was always
emptied. It had no impact though, because the notifications
were not delivered.

It is used to update the EMF model when attributes are
removed from the Rodin database. In particular, this is
the case of the boolean attributes, which are generally
deleted when they are set to 'false'.

It seems to have fixed a bug with the 'theorem' attribute
that was not updated after a click in the Rodin editor
(when changing from 'theorem' to 'not theorem').