Fix Problem occurred while retrieving the generated attribute
When testing on Windows, I have noticed several occurrences of this
issue in the Eclipse log.  This can happen because of a concurrent
access to the database while cleaning a project.

This is only a partial cure, as the element can still disappear
between the existence test and the reading of the attribute,
but this should happen much less often.