Fix invalid API restriction for IRodinElement and co
The purpose of the @noimplement annotation is to allow the plug-in
implementer to later enrich the interface while keeping backward
compatibility.

However, as soon as we allow people to extend the RodinElement class,
there is no such liberty anymore: we must not add any method to its
parent interfaces.

Hence the removal of the annotation and the corresponding problem
filters.

No functional change.