Fix class initialization order issue in IPosition
With the change to Java 8, an existing issue with class initialization
order has been uncovered.  Depending on whether IPosition or Position
starts being initialized first, a NullPointerException can be thrown by
the code.  This depends on the order in which the static fields NO_INTS
and ROOT are initialzed.

To fix this, we remove the NO_INTS static field and adapt the code
consequently.  The new version is actually a bit more performant (will
not create a new instance for root nodes when deserializing from a
String).