Remove useless attribute
The tableValues attribute is not necessary.  It can be replaced by
calling the values() method of the HashTable.

The modification does not change anything because all values are
different by construction (they all have different symbols).  There is
no risk to have two equal values in the result and we can thus replace
the set by a collection.

The corresponding unit tests needed to be updated, because the new
result is not a proper set anymore (which was not guaranteed by the API,
anyhow). We therefore create an intermediate set to get a proper
comparison.

No functional change.